requirements/requirements.txt
discord.py
discordhealthcheck
requests
lxml
pytz
python_weather
pyyaml
psycopg[binary]
